Moves ModelManager to the root model folder
