+++ core/src/main/se/tevej/game/GameManager.java
@@ -4,7 +4,7 @@ import com.badlogic.gdx.ApplicationAdapter;
 import com.badlogic.gdx.Gdx;
 
 import main.se.tevej.game.controller.ControllerManager;
-import main.se.tevej.game.model.ashley.ModelManager;
+import main.se.tevej.game.model.ModelManager;
 import main.se.tevej.game.view.ViewManager;
 
 public class GameManager extends ApplicationAdapter {
