+++ core/src/main/se/tevej/game/model/entities/WorldEntity.java
@@ -5,7 +5,7 @@ import java.util.List;
 
 import com.badlogic.ashley.core.Entity;
 
-import main.se.tevej.game.model.ashley.ModelManager;
+import main.se.tevej.game.model.ModelManager;
 import main.se.tevej.game.model.components.PositionComponent;
 import main.se.tevej.game.model.components.TileComponent;
 import main.se.tevej.game.model.components.WorldComponent;
