+++ core/src/test/java/TestWorldComponent.java
@@ -3,7 +3,7 @@ package test.java;
 import static org.junit.Assert.assertTrue;
 
 import main.se.tevej.game.Options;
-import main.se.tevej.game.model.ashley.ModelManager;
+import main.se.tevej.game.model.ModelManager;
 import main.se.tevej.game.model.components.PositionComponent;
 import main.se.tevej.game.model.components.WorldComponent;
 import main.se.tevej.game.model.entities.WorldEntity;
